added new CLI, improved sweep support & refactoring

+++ lynx/envs/wrappers/jumanji/transformers/puzzle.py
@@ -12,10 +12,10 @@ class PuzzleAgentViewTransformer:
     def __call__(self, observation: SlidingTilePuzzleObservation) -> chex.Array:
         puzzle = observation.puzzle.ravel()
         empty = observation.empty_tile_position.ravel()
-        return jnp.concatenate([puzzle, empty])
+        return jnp.concatenate([puzzle, empty]).astype(jnp.float32)
 
     def get_observation_spec(self, env: Environment) -> Spec:
         puzzle_shape = jnp.asarray(env.observation_spec.puzzle.shape)
         empty_tile_shape = jnp.asarray(env.observation_spec.empty_tile_position.shape)
         agent_view_size = puzzle_shape.prod() + empty_tile_shape.prod()
-        return Array(shape=(agent_view_size,), dtype=float)
+        return Array(shape=(agent_view_size,), dtype=jnp.float32)
